Feature/fcash erc20 (#6)
* upgrading to solidity 8

* initial version of wrapped fCash

* updating environment config

* deployment tests

* adding minting tests

* updated mint and redeem tests

* adding test for contract receivers

* adding trading functions

* adding wrapped fcash interface

* refactoring a bit to add different redemption methods

* adding first cut of Set Token exchange adapter

* getting rid of fCash trading

* adding fUSDC testing

* adding first cut of pre issue hook

* support fETH minting

* adding reentrancy guard for ETH

* renaming file

* refactoring erc20 contract

* some tests passing

* tests passing on kovan

* trade module tests passing

* initial gas estimates

* updating wrapper code with tests

* deployed and verified on kovan

+++ contracts/abstract/NotionalV2BaseLiquidator.sol
@@ -128,7 +128,7 @@ abstract contract NotionalV2BaseLiquidator {
 
         // Will withdraw entire cash balance. Don't redeem local currency here because it has been flash
         // borrowed and we need to redeem the entire balance to underlying for the flash loan repayment.
-        _redeemAndWithdraw(localCurrency, uint96(netNTokens), false);
+        _redeemAndWithdraw(localCurrency, uint96(int96(netNTokens)), false);
     }
 
     function _liquidateCollateral(
@@ -167,7 +167,7 @@ abstract contract NotionalV2BaseLiquidator {
         );
 
         // Redeem to underlying for collateral because it needs to be traded on the DEX
-        _redeemAndWithdraw(collateralCurrency, uint96(collateralNTokens), true);
+        _redeemAndWithdraw(collateralCurrency, uint96(int96(collateralNTokens)), true);
         SafeToken.redeemCTokenEntireBalance(collateralAddress);
 
         // Wrap everything to WETH for trading
